+++ Classes/Tsunami.java
@@ -2,7 +2,7 @@ package Classes;
 public class Tsunami extends NaturalCalamities{
     
     public void modifyWaterAmount(){
-        Ground.waterAmount+=0.01*duration*destructionLevel;
+        Ground.waterAmount+=(float)(0.01*duration*destructionLevel);
         World.humidity+=0.01*duration*destructionLevel;
         if (Ground.waterAmount>100)
             Ground.waterAmount=100;
